+++ src/space/errors.ts
@@ -11,7 +11,6 @@
 export class HumanReadableError extends Error {
   constructor(message: string) {
     super(message)
-    this.message = message
     this.name = 'HumanReadableError'
   }
 }